TODO Finish recursive sudoku solving algorithm and potentially change to use an iterative stack-based implementation.
Bugfix issues with selenium suddenly not finding the content I need. Maybe there was a change in the NYT website?

Overall, I'm happy to leave this project as is. I mostly did this as an exercise to keep my coding skills sharp and to learn how to webscrape for future projects.